Add IBM PS/2 protocol information from computer-engineering.org.
Copyright 1998-2004, Adam Chapweske, with the attached license:

``All information within this article is provided "as is" and
without any express or implied warranties, including, without
limitation, the implied warranties of merchantibility and fitness
for a particular purpose.

This article is protected under copyright law. This document may
be copied only if the source, author, date, and legal information
is included. ''